requirements.txt
Pillow>=10.0
numpy>=1.24
scipy>=1.11
customtkinter>=5.2
